requirements.txt
# Project 01: Classical Lane Detection Dependencies

# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.4.0
PyYAML>=6.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Code quality
black>=22.0.0
flake8>=4.0.0

# Analysis and reporting
pandas>=1.4.0
seaborn>=0.11.0
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0
Pillow>=9.0.0
